requirements.txt
astroid==2.5.1
bidict==0.21.2
blinker==1.4
click==7.1.2
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
gunicorn==20.0.4
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.6
PyJWT==2.0.1
pylint==2.7.2
python-engineio==4.0.1
python-socketio==5.1.0
six==1.15.0
SQLAlchemy==1.4.2
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
